Back-end/BE/requirements.txt
Django==4.2.3
django-cors-headers==4.3.0
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2

Pillow==10.1.0